ref: Move the current table to old_ui
The beginning of #328, the current table is deprecated and being rethought/rewritten.
