+++ src/pages/RepoPage/CommitsTab/CommitsTable/CommitsTable.jsx
@@ -1,10 +1,10 @@
 import PropTypes from 'prop-types'
 import { useParams } from 'react-router-dom'
 
+import Table from 'old_ui/Table'
 import { useCommits } from 'services/commits'
 import Button from 'ui/Button'
 import Spinner from 'ui/Spinner'
-import Table from 'ui/Table'
 import TotalsNumber from 'ui/TotalsNumber'
 
 import CIStatus from './CIStatus'
