+++ src/pages/RepoPage/FlagsTab/subroute/FlagsTable/FlagsTable.jsx
@@ -2,13 +2,13 @@ import PropTypes from 'prop-types'
 import { useMemo, useState } from 'react'
 import { useParams } from 'react-router-dom'
 
+import Table from 'old_ui/Table'
 import { useRepoConfig } from 'services/repo/useRepoConfig'
 import { determineProgressColor } from 'shared/utils/determineProgressColor'
 import Button from 'ui/Button'
 import CoverageProgress from 'ui/CoverageProgress'
 import Icon from 'ui/Icon'
 import Spinner from 'ui/Spinner'
-import Table from 'ui/Table'
 
 import DeleteFlagModal from './DeleteFlagModal'
 import useRepoFlagsTable from './hooks'
