+++ src/shared/ListRepo/ReposTable/ReposTable.jsx
@@ -2,13 +2,13 @@ import isEmpty from 'lodash/isEmpty'
 import PropTypes from 'prop-types'
 import { useContext } from 'react'
 
+import Table from 'old_ui/Table'
 import { useRepos } from 'services/repos'
 import { useOwner, useUser } from 'services/user'
 import { ActiveContext } from 'shared/context'
 import { formatTimeToNow } from 'shared/utils/dates'
 import Button from 'ui/Button'
 import Spinner from 'ui/Spinner'
-import Table from 'ui/Table'
 import TotalsNumber from 'ui/TotalsNumber'
 
 import InactiveRepo from './InactiveRepo'
